Add ability to add a new Activity

+++ A3D.Library.Services/ActivityService.cs
@@ -5,6 +5,8 @@ using Microsoft.EntityFrameworkCore;
 using System.Collections.Generic;
 using System;
 using System.Linq;
+using System.Security;
+using A3D.Library.Models.LookUp;
 
 namespace A3D.Library.Services
 {
@@ -24,8 +26,38 @@ namespace A3D.Library.Services
             return this.activityRepository.Create(item);
         }
 
+        /// <summary>
+        /// Delete an Activity by id.
+        /// 
+        /// PERMISSIONS
+        ///   Only allow Activity creator to delete it.
+        ///   TODO allow admin in the future?
+        /// </summary>
+        /// <param name="context"></param>
+        /// <param name="id"></param>
         public void DeleteById(ApplicationContext context, int id)
         {
+            // If there's no current user, don't allow
+            if (context.CurrentUser == null)
+            {
+                throw new SecurityException($"User NULL is unauthorized to delete Activity id {id}");
+            }
+
+            var activity = this.activityRepository.GetById(id);
+            
+            // If the item doesn't exist, treat it as if the item has been deleted
+            if (activity == null)
+            {
+                return;
+            }
+
+            // TODO allow Admin 
+            // If the current user is not the item creator, don't allow
+            if (activity.CreatorId != context.CurrentUser.Id)
+            {
+                throw new SecurityException($"User {context.CurrentUser.Id} is unauthorized to delete Activity id {id}");
+            }
+            
             try
             {
                 this.activityRepository.DeleteById(id);
@@ -42,9 +74,30 @@ namespace A3D.Library.Services
             }
         }
 
+        /// <summary>
+        /// Returns an Activity by id.
+        /// 
+        /// PERMISSIONS
+        ///   Allow anybody to see all public Activities.
+        ///   Only allow Activity creator to see private Activities.
+        /// </summary>
+        /// <param name="context"></param>
+        /// <param name="id"></param>
+        /// <returns></returns>
         public Activity GetById(ApplicationContext context, int id)
         {
-            return this.activityRepository.GetById(id);
+            var activity = this.activityRepository.GetById(id);
+
+            if (activity != null)
+            {
+                if (activity.PrivacyId == ActivityPrivacy.Private.Id
+                    && activity.CreatorId != context.CurrentUser.Id)
+                {
+                    throw new SecurityException($"User {context.CurrentUser.Id} is unauthorized to view Activity id {id}");
+                }
+            }
+
+            return activity;
         }
 
         public IEnumerable<Activity> GetByCreatorId(ApplicationContext context, string creatorId)
@@ -61,7 +114,26 @@ namespace A3D.Library.Services
 
         public void Update(ApplicationContext context, Activity item)
         {
+            // If there's no current user, don't allow
+            if (context.CurrentUser == null)
+            {
+                throw new SecurityException($"User NULL is unauthorized to delete Activity id {item.Id}");
+            }
+
             var existingItem = this.activityRepository.GetById(item.Id);
+            
+            // If the item doesn't exist, throw an exception
+            if (existingItem == null)
+            {
+                throw new NullReferenceException($"Activity id {item.Id} does not exist.");
+            }
+
+            // TODO allow Admin 
+            // If the current user is not the item creator, don't allow
+            if (existingItem.CreatorId != context.CurrentUser.Id)
+            {
+                throw new SecurityException($"User {context.CurrentUser.Id} is unauthorized to update Activity id {item.Id}.");
+            }
 
             // TODO make this work with PATCH
             existingItem.Name = item.Name;
